requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.12.1
autogen-agentchat==0.7.5
autogen-core==0.7.5
autogen-ext==0.7.5
certifi==2026.2.25
charset-normalizer==3.4.5
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
jiter==0.13.0
jsonref==1.1.0
openai==2.26.0
opentelemetry-api==1.40.0
pillow==12.1.1
protobuf==5.29.6
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
regex==2026.2.28
requests==2.32.5
sniffio==1.3.1
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
zipp==3.23.0
